// src/types.ts
export type RnCheckBox = { label: string, value: string, isSelected: boolean }

export interface RnCheckboxGroupProps extends TouchableOpacityProps {
    value: string,
    label?: string,
    fields: RnCheckBox[],
    disable: boolean,
    required: boolean,
    onChange: Function
}